Fix to use useContractWrite and refactor
