+++ hooks/quest/useKeywordSubmit.ts
@@ -1,28 +1,49 @@
 import { ChangeEventHandler, useCallback, useState } from 'react'
-import { useKamonNFT } from '@/hooks/useKamonNFT'
 import { useToast } from '@/hooks/useToast'
+import { useContractWrite, useNetwork } from 'wagmi'
+import contractInterface from '@/abis/PodCastNFT.json'
+import { getContractAddress } from '@/utils/contractAddress'
+import { getErrorMessageFromRaw } from '@/utils/handleError'
 
 export const useKeywordSubmit = () => {
-  const { contract } = useKamonNFT()
   const { toast } = useToast()
-
+  const { activeChain } = useNetwork()
+  const kamonNFT = getContractAddress({
+    name: 'kamonNFT',
+    chainId: activeChain?.id
+  })
   const [keyword, setKeyword] = useState<string>('')
-  const [isLoading, setIsLoading] = useState(false)
+  const { write: submit, isLoading: isSubmitting } = useContractWrite(
+    {
+      addressOrName: kamonNFT,
+      contractInterface: contractInterface.abi
+    },
+    'checkAnswer',
+    {
+      args: [keyword],
+      onError(error) {
+        toast({
+          title: 'Error',
+          description: handleErrorMessage(error),
+          status: 'error'
+        })
+      },
+      onSuccess() {
+        toast({
+          title: 'Submitted',
+          description: 'Answer was correct.\nYou earn 100 points',
+          status: 'success'
+        })
+      }
+    }
+  )
 
   const inputChange: ChangeEventHandler<HTMLInputElement> = (event) => {
     setKeyword(event.target.value)
   }
 
-  const getErrorMessageFromRaw = (message: string): string => {
-    const regexp =
-      /Error: VM Exception while processing transaction: reverted with reason string '([a-zA-Z :]+)'/
-    const result = regexp.exec(message)
-    return result && result.length > 1 ? result[1] : ''
-  }
-
-  const handleErrorMessage = useCallback((raw: string): string => {
-    const message = getErrorMessageFromRaw(raw)
-    switch (message) {
+  const handleErrorMessage = useCallback((error: Error): string => {
+    switch (getErrorMessageFromRaw(error)) {
       case 'MUST BE HOLDER':
         return 'Only token holders may answer.'
       case 'WRONG ANSWER':
@@ -34,26 +55,5 @@ export const useKeywordSubmit = () => {
     }
   }, [])
 
-  const submit = useCallback(async () => {
-    setIsLoading(true)
-    await contract
-      .checkAnswer(keyword)
-      .then(() =>
-        toast({
-          title: 'Submitted',
-          description: 'Answer was correct.\nYou earn $100 henkaku',
-          status: 'success'
-        })
-      )
-      .catch((e) => {
-        toast({
-          title: 'Error',
-          description: handleErrorMessage(e.data.message),
-          status: 'error'
-        })
-      })
-      .finally(() => setIsLoading(false))
-  }, [contract, handleErrorMessage, keyword, toast])
-
-  return { keyword, inputChange, submit, isLoading }
+  return { keyword, inputChange, submit, isSubmitting }
 }
