+++ tsconfig.json
@@ -18,7 +18,8 @@
     "paths": {
       "@/components/*": ["components/*"],
       "@/hooks/*": ["hooks/*"],
-      "@/contracts/*": ["contracts/*"]
+      "@/abis/*": ["abis/*"],
+      "@/utils/*": ["utils/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
